% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_helper.R
\name{api_call}
\alias{api_call}
\alias{api_calls}
\title{API-Request ausführen}
\usage{
api_call(endpoint, ...)

api_calls(endpoint, jahr = NULL, code = NULL, name = NULL)
}
\arguments{
\item{endpoint}{URL-Endpunkt}

\item{...}{Zusätzliche Komponenten `query parameter = value`. Z.B.: code = 111,
name = "Züri" oder jahr = 1991.}
}
\value{
JSON-Response als Dataframe oder Liste.
}
\description{
Funktion zum Abfragen sämtlicher API Endpunkte
\itemize{
\item \code{api_call()}: Für einzellne abfragen.
\item \code{api_calls()}: Für deb Fall dass \code{jahr}, \code{code} oder
 \code{name} nicht \code{NULL} ist.
}
}
\keyword{internal}
