% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_helper.R
\name{call_and_parse}
\alias{call_and_parse}
\title{API-Request senden und parsen}
\usage{
call_and_parse(request)
}
\arguments{
\item{request}{Request-Objekt von `generate_api_request()`.}
}
\value{
JSON-Response als Dataframe oder Liste.
}
\description{
API-Request senden und JSON-Response parsen.
}
\keyword{internal}
