% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_helper.R
\name{check_input_param}
\alias{check_input_param}
\title{Fehler abfangen}
\usage{
check_input_param(jahr = NULL, code = NULL, name = NULL)
}
\arguments{
\item{jahr}{Jahr von interesse.}

\item{code}{Code der Gemeinde, Region oder Raumplanungsregion.}

\item{name}{Name der Gemeinde, Region oder Raumplanungsregion.}
}
\value{
Kein Rückgabeobjekt.
}
\description{
In dieser Funktion werden mögliche fehlerhafte Eingaben
abgefangen.
}
\keyword{internal}
