% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_helper.R
\name{generate_api_request}
\alias{generate_api_request}
\title{Request-Objekt erstellen}
\usage{
generate_api_request(endpoint, ...)
}
\arguments{
\item{endpoint}{URL-Endpunkt}

\item{...}{Zusätzliche Komponenten `query parameter = value`. Z.B.: code = 111,
name = "Züri" oder jahr = 1991.}
}
\value{
Ein Request-Objekt.
}
\description{
Request-Objekt erstellen mit angegebener URL-Endpunkt und
gegebenenfalls Filtern.
}
\keyword{internal}
