% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carrier_wrapper.R
\name{get_gebiete}
\alias{get_gebiete}
\title{Alle Gebiete abrufen}
\usage{
get_gebiete()
}
\value{
Dataframe mit Gebietstyp-Code und Name sowie Code und Name der Gebiete.
}
\description{
Ruft alle Gebiete (Kanton, Bezirke, Gemeinden und
Raumplanungsregionen) des Kantons Zürich (Stand heute) ab.
}
\examples{
\dontrun{
# Alle Gebiete abrufen
get_gebiete()
}

}
