% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carrier_wrapper.R
\name{get_gebietstypen}
\alias{get_gebietstypen}
\title{Alle Gebietstypen abrufen}
\usage{
get_gebietstypen()
}
\value{
Dataframe mit Gebietstyp-Code und Name.
}
\description{
Ruft alle Gebietstypen auf.
}
\examples{
\dontrun{
# Alle Gebietstypen abrufen
get_gebietstypen()
}

}
