% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carrier_wrapper.R
\name{get_gemeindemutationen}
\alias{get_gemeindemutationen}
\title{Alle Gemeindemutationen abrufen}
\usage{
get_gemeindemutationen()
}
\value{
Ein Dataframe mit dem Mutationstyp, dem alten Gemeindecode und Namen
 sowie dem neuen Gemeindecode und Namen, so wie das Mutationsdatum.
}
\description{
Ruft alle Gemeindemutationen seit 1990 ab.
}
\examples{
\dontrun{
# Alle Gemeindemutationen abrufen
get_gemeindemutationen()
}

}
