% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carrier_wrapper.R
\name{get_gemeinden}
\alias{get_gemeinden}
\title{Gemeinden abrufen}
\usage{
get_gemeinden(auswahl = FALSE, ..., gemeinde_code = NULL, gemeinde_name = NULL)
}
\arguments{
\item{auswahl}{Boolean, welcher beschreibt, ob eine Auswahl über die gefundenen
Namen gemacht werden soll oder nicht. Verwendungszweck: beim Harmonisieren
von Namen.}

\item{...}{Unten gelistete Filterargumente:}

\item{gemeinde_code}{Filter nach Gemeindecode. Muss eine ganze positive Zahl
sein. Ein Vektor aus verschiedenen Gemeindecodes ist ebenfalls zulässig.}

\item{gemeinde_name}{Filter nach Gemeinde Name. Muss vom Typ `character` sein.
Ein Vektor aus verschiedenen Gemeinde Namen ist ebenfalls zulässig.}
}
\value{
Dataframe mit Gebietstyp-Code, Offiziellem Code der Gemeinde
(BFS-Nummer) und Offiziellem Namen der Gemeinde (BFS-Name).
}
\description{
Ermöglicht das Laden aller Gemeinden mit oder ohne Filter.
}
\details{
Es kann nur \code{gemeinde_code} oder \code{gemeinde_name} angegeben
werden, aber nicht beide Argumente gemeinsam.\cr\cr
Sämtliche angegebenen Filteroptionen müssen zu einem Treffer führen.
}
\examples{
\dontrun{
# Alle Gemeinden abrufen
get_gemeinden()

# Filtern nach einem oder mehreren Codes
get_gemeinden(gemeinde_code = 111)
get_gemeinden(gemeinde_code = c(111, 112))
get_gemeinden(gemeinde_code = c(111:120))

# Filtern nach einem oder mehreren Namen
get_gemeinden(auswahl = TRUE, gemeinde_name = "am Albis")
get_gemeinden(gemeinde_name = c("am Albis", "Züri"))
}

}
