% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carrier_wrapper.R
\name{get_gemeindenhist}
\alias{get_gemeindenhist}
\title{Jahresstände von Gemeinden abrufen}
\usage{
get_gemeindenhist(..., jahr = NULL, gemeinde_code = NULL)
}
\arguments{
\item{...}{Unten gelistete Filterargumente:}

\item{jahr}{Filter nach spezifischem Jahr. Muss eine ganze Zahl sein, welche
zwischen 1990 und dem heutigen Jahr liegt.}

\item{gemeinde_code}{Filter nach Gemeindecode. Muss eine ganze positive Zahl
sein. Ein Vektor aus verschiedenen Gemeindecodes ist ebenfalls zulässig.}
}
\value{
Dataframe mit Gebietstyp-Code, offiziellem Code der Gemeinde
(BFS-Nummer), offiziellem Namen der Gemeinde (BFS-Name) und Jahr des
Jahresbestands.
}
\description{
Jahresstände aller Gemeinden abrufen, ab 1990 mit oder ohne
Filter.
}
\details{
Alle Kombinationen von \code{jahr} und \code{gemeinde_code} sind
 möglich.
}
\examples{
\dontrun{
# Alle Jahresstände von Gemeinden abrufen
get_gemeindenhist()

# Filtern nach einem oder mehreren Jahren
get_gemeindenhist(jahr = 2026)
get_gemeindenhist(jahr = c(2021:2026))

# Filtern nach einem oder mehreren Gemeindecodes
get_gemeindenhist(gemeinde_code = 111)
get_gemeindenhist(gemeinde_code = c(111, 113))

# Filtern nach Jahr und Gemeindecode
get_gemeindenhist(jahr = 2021, gemeinde_code = 111)
get_gemeindenhist(jahr = c(2021:2026), gemeinde_code = c(111, 113))
}

}
