% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carrier_wrapper.R
\name{get_health}
\alias{get_health}
\title{Health-Check der API}
\usage{
get_health()
}
\value{
Liste mit den folgenden Komponenten:
\describe{
  \item{status}{Aktueller Status der API.}
  \item{timestamp}{Zeitpunkt der Abfrage.}
  \item{data_available}{Boolean, der angibt, ob Daten verfügbar sind.}
}
}
\description{
Fragt die Health-Parameter der API ab.
}
\examples{
\dontrun{
# Health-Check ausfuehren
get_health()
}

}
