% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser_helper.R
\name{parse_to_df}
\alias{parse_to_df}
\title{Data frame Parser}
\usage{
parse_to_df(list)
}
\arguments{
\item{list}{Eine aus einer API-Abfrage stammende Liste.}
}
\value{
Ein Datenframe mit den geparsten Ergebnissen.
}
\description{
Konvertiert eine Liste aus einer API-Abfrage in einen Datenframe.
}
\keyword{internal}
