% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/name_selector.R
\name{select_name}
\alias{select_name}
\title{Namen bei Mehrfachtreffern auswaehlen}
\usage{
select_name(list, selection, search_term = NULL)
}
\arguments{
\item{list}{Eine aus einer API-Abfrage stammende Liste.}

\item{selection}{Boolean, ob eine Auswahl erfolgen soll.}

\item{search_term}{Optional: Der verwendete Suchbegriff für die Ausgabe.}
}
\value{
Trefferliste mit dem ausgewählten Namen im Original-Wrapper.
}
\description{
Namen bei Mehrfachtreffern auswaehlen
}
\keyword{internal}
