% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carrier_wrapper.R
\name{get_bezirke}
\alias{get_bezirke}
\title{Bezirke abrufen}
\usage{
get_bezirke(..., bezirk_code = NULL, bezirk_name = NULL)
}
\arguments{
\item{...}{Unten gelistete Filterargumente:}

\item{bezirk_code}{Filter nach Bezirkcode. Muss eine ganze positive Zahl sein.
Ein Vektor aus verschiedenen Bezirkcodes ist ebenfalls zulässig.}

\item{bezirk_name}{Filter nach Bezirk Name. Muss vom Typ `character` sein.
Ein Vektor aus verschiedenen Bezirk Namen ist ebenfalls zulässig.}
}
\value{
Dataframe mit Gebietstyp-Code, Offizieller Code des Bezirks
(BFS-Nummer) und Offizieller Name des Bezirks (BFS-Name).
}
\description{
Ermöglicht das Laden aller Bezirke mit oder ohne Filter.
}
\details{
Es kann nur \code{bezirk_code} oder \code{bezirk_name} angegeben
werden, aber nicht beide Argumente gemeinsam.\cr\cr
Sämtliche angegebenen Filteroptionen müssen zu einem Treffer führen.
}
\examples{
\dontrun{
# Alle Bezirke abrufen
get_bezirke()

# Filtern nach einem oder mehreren Codes
get_bezirke(bezirk_code = 111)
get_bezirke(bezirk_code = c(111, 112))
get_bezirke(bezirk_code = c(110:112))

# Filtern nach einem oder mehreren Namen
get_bezirke(bezirk_name = "Andeldingen")
get_bezirke(bezirk_name = c("Andeldingen", "Züri"))
}
}
