% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carrier_wrapper.R
\name{get_gemeindezuweisungen}
\alias{get_gemeindezuweisungen}
\title{Gemeindezuweisungen abrufen}
\usage{
get_gemeindezuweisungen(..., gemeinde_code = NULL)
}
\arguments{
\item{...}{Unten gelistete Filterargumente:}

\item{gemeinde_code}{Filter nach Gemeindecode. Muss eine ganze positive Zahl
sein. Ein Vektor aus verschiedenen Gemeindecodes ist ebenfalls zulässig.}
}
\value{
Dataframe mit offiziellem Code der Gemeinde (BFS-Nummer),
offiziellem Namen der Gemeinde (BFS-Name), offiziellem Code des Bezirks
(BFS-Nummer), offiziellem Name des Bezirks (BFS-Name), offiziellem Code der
Raumplanungsregion (Vergabe durch ARE/Kantone) und offiziellem Name der
Raumplanungsregion (Vergabe durch ARE/Kantone).
}
\description{
Ruft die aktuelle Zuweisung aller Gemeinden zu einem Bezirk und
einer Raumplanungsregion ab. Diese Abfrage ist mit oder ohne Filter verfügbar.
}
\examples{
\dontrun{
# Alle Gemeindezuweisungen abrufen
get_gemeindezuweisungen()

# Filtern nach einem oder mehreren Gemeindecodes
get_gemeindezuweisungen(gemeinde_code = 111)
get_gemeindezuweisungen(gemeinde_code = c(111, 113))
}

}
