% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carrier_wrapper.R
\name{get_raumplanungsregionen}
\alias{get_raumplanungsregionen}
\title{Raumplanungsregionen abrufen}
\usage{
get_raumplanungsregionen(
  auswahl = FALSE,
  ...,
  raumplanungsregion_code = NULL,
  raumplanungsregion_name = NULL
)
}
\arguments{
\item{auswahl}{Boolean, welcher beschreibt, ob eine Auswahl über die gefundenen
Namen gemacht werden soll oder nicht. Verwendungszweck: beim Harmonisieren
von Namen.}

\item{...}{Unten gelistete Filterargumente:}

\item{raumplanungsregion_code}{Filter nach Raumplanungsregioncode. Muss eine
ganze positive Zahl sein. Ein Vektor aus verschiedenen
Raumplanungsregioncodes ist ebenfalls zulässig.}

\item{raumplanungsregion_name}{Filter nach Rraumplanungsregion Name. Muss
vom Typ `character` sein. Ein Vektor aus verschiedenen Raumplanungsregion
Namen ist ebenfalls zulässig.}
}
\value{
Dataframe mit Gebietstyp-Code, Offizieller Code der
Raumplanungsregion (Vergabe durch ARE/Kantone) und Offizieller Name der
Raumplanungsregion (Vergabe durch ARE/Kantone).
}
\description{
Ermöglicht das Laden aller Raumplanungsregionen mit oder ohne
Filter.
}
\details{
Es kann nur \code{raumplanungsregion_code} oder
\code{raumplanungsregion_name} angegeben werden, aber nicht beide Argumente
gemeinsam.\cr\cr
Sämtliche angegebenen Filteroptionen müssen zu einem Treffer führen.
}
\examples{
\dontrun{
# Raumplanungsregionen abrufen
get_raumplanungsregionen()

# Filtern nach einem oder mehreren Codes
get_raumplanungsregionen(raumplanungsregion_code = 101)
get_raumplanungsregionen(raumplanungsregion_code = c(101, 103))
get_raumplanungsregionen(raumplanungsregion_code = c(101:103))

# Filtern nach einem oder mehreren Namen
get_raumplanungsregionen(auswahl = TRUE, raumplanungsregion_name = "tal")
get_raumplanungsregionen(raumplanungsregion_name = c("tal", "Züri"))
}

}
