% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser_helper.R
\name{parse_to_df}
\alias{parse_to_df}
\title{Data frame Parser}
\usage{
parse_to_df(list)
}
\arguments{
\item{list}{Eine aus einer API-Abfrage stammende Liste.}
}
\value{
Daten frame.
}
\description{
Konvertiert eine Liste in einen Datenframe
}
\keyword{internal}
