% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser_helper.R
\name{remove_gemeinden}
\alias{remove_gemeinden}
\title{Entferne "Gemeinden"-Liste}
\usage{
remove_gemeinden(list)
}
\arguments{
\item{list}{Eine aus einer API-Abfrage stammende Liste.}
}
\value{
Die Eingabeliste ohne das Element "Gemeinden".
}
\description{
Bei den Bezirken und Raumplanungsregionen wird im API-Abruf eine
Liste der zugehörigen Gemeinden zurückgegeben. Diese Liste wird mit dieser
Funktion entfernt.
}
\keyword{internal}
