% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/name_selector.R
\name{select_name}
\alias{select_name}
\title{Namen bei Mehrfachtreffern auswaehlen}
\usage{
select_name(list)
}
\arguments{
\item{list}{Eine aus einer API-Abfrage stammende Liste.}
}
\value{
Trefferliste mit dem ausgewählten Namen und ohne den Suchnamen.
}
\description{
Namen bei Mehrfachtreffern auswaehlen
}
\details{
Diese Funktion erlaubt es zu identifizieren, ob eine Suche nach
Namen mehrere Treffer geliefert hat. Falls ja, muessen Nutzende via Konsole
ein Namen auswäaelen.
}
\keyword{internal}
